requirements.txt
pandas
openpyxl
phonenumbers
python-dotenv
google-generativeai
streamlit
tabulate
fuzzywuzzy
python-Levenshtein
